- mmp_pread exported - mmp_pread and mmp_pwrite offset parameter changed to off_t

+++ mmp_files.c
@@ -63,14 +63,14 @@ int mmp_close(int fd)
 }
 
 /** \todo missing unittest */
-int mmp_pread(int fd, const void *buf, size_t nbyte, long offset)
+int mmp_pread(int fd, const void *buf, size_t nbyte, off_t offset)
 {
 #ifndef _WIN32
     return pread(fd, buf, nbyte, offset);
 #else
     long pos, ret;
     pos = _tell(fd);
-    _lseek(fd, offset, SEEK_SET);
+    _lseek(fd, (long)offset, SEEK_SET);
     ret = _read(fd, buf, nbyte);
     _lseek(fd, pos, SEEK_SET);
     return ret;
@@ -78,14 +78,14 @@ int mmp_pread(int fd, const void *buf, size_t nbyte, long offset)
 }
 
 /** \todo missing unittest */
-int mmp_pwrite(int fd, const void *buf, size_t nbyte, long offset)
+int mmp_pwrite(int fd, const void *buf, size_t nbyte, off_t offset)
 {
 #ifndef _WIN32
     return pwrite(fd, buf, nbyte, offset);
 #else
     long pos, ret;
     pos = _tell(fd);
-    _lseek(fd, offset, SEEK_SET);
+    _lseek(fd, (long)offset, SEEK_SET);
     ret = _write(fd, buf, nbyte);
     _lseek(fd, pos, SEEK_SET);
     return ret;
